requirements.txt
pyTelegramBotAPI==4.15.2
requests==2.31.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Pillow==10.0.1